PR review suggestions - transfer code to ReportStreamLabOrderSender
