--- app/src/main/java/gov/hhs/cdc/trustedintermediary/wrappers/ClientConnection.java
@@ -1,8 +1,0 @@
-package gov.hhs.cdc.trustedintermediary.wrappers;
-
-public interface ClientConnection {
-
-    void sendRequestBody(String json, String clientToken);
-
-    String requestToken();
-}
